refactor:add Interface for CustomMailSender
