+++ gg-utils/src/main/java/gg/utils/sns/MailSender.java
@@ -1,0 +1,6 @@
+package gg.utils.sns;
+
+public interface MailSender {
+
+	void send(String emailTo, String subject, String text);
+}
